# PERF/ENH: allow Generator in sampling methods