--- a/pandas/_typing.py
+++ b/pandas/_typing.py
@@ -122,7 +122,9 @@ Ordered = Optional[bool]
 JSONSerializable = Optional[Union[PythonScalar, List, Dict]]
 Frequency = Union[str, "DateOffset"]
 Axes = Collection[Any]
-RandomState = Union[int, ArrayLike, np.random.Generator, np.random.RandomState]
+RandomState = Union[
+    int, ArrayLike, np.random.Generator, np.random.BitGenerator, np.random.RandomState
+]
 
 # dtypes
 NpDtype = Union[str, np.dtype]