--- a/pandas/_typing.py
+++ b/pandas/_typing.py
@@ -126,6 +126,7 @@ JSONSerializable = Optional[Union[PythonScalar, List, Dict]]
 Frequency = Union[str, "DateOffset"]
 Axes = Collection[Any]
 
+# BitGenerator isn't exposed until 1.18
 if TYPE_CHECKING:
     RandomState = Union[
         int,
@@ -134,8 +135,6 @@ if TYPE_CHECKING:
         np.random.BitGenerator,
         np.random.RandomState,
     ]
-else:
-    RandomState = Union[int, ArrayLike, np.random.Generator, np.random.RandomState]
 
 # dtypes
 NpDtype = Union[str, np.dtype]