--- a/pandas/core/common.py
+++ b/pandas/core/common.py
@@ -32,7 +32,6 @@ from pandas._typing import (
     AnyArrayLike,
     ArrayLike,
     NpDtype,
-    RandomState,
     Scalar,
     T,
 )
@@ -54,6 +53,9 @@ from pandas.core.dtypes.inference import iterable_not_string
 from pandas.core.dtypes.missing import isna
 
 if TYPE_CHECKING:
+    # Includes BitGenerator, which only exists >= 1.18
+    from pandas._typing import RandomState
+
     from pandas import Index